Revert manager just use multiple classes
