+++ config/services.yml
@@ -8,7 +8,7 @@ services:
             - '@auth'
             - '@config'
             - '@controller.helper'
-            - '@phpbb.ideas.manager'
+            - '@phpbb.ideas.idea'
             - '@language'
             - '@phpbb.ideas.linkhelper'
             - '@template'
@@ -37,7 +37,6 @@ services:
             - '@auth'
             - '@config'
             - '@controller.helper'
-            - '@phpbb.ideas.manager'
             - '@language'
             - '@phpbb.ideas.linkhelper'
             - '@pagination'
@@ -50,10 +49,14 @@ services:
     phpbb.ideas.index_controller:
         class: phpbb\ideas\controller\index_controller
         parent: phpbb.ideas.controller.base
+        calls:
+            - [get_entity, ['@phpbb.ideas.ideas']]
 
     phpbb.ideas.list_controller:
         class: phpbb\ideas\controller\list_controller
         parent: phpbb.ideas.controller.base
+        calls:
+            - [get_entity, ['@phpbb.ideas.ideas']]
 
     phpbb.ideas.post_controller:
         class: phpbb\ideas\controller\post_controller
@@ -62,19 +65,16 @@ services:
     phpbb.ideas.idea_controller:
         class: phpbb\ideas\controller\idea_controller
         parent: phpbb.ideas.controller.base
+        calls:
+            - [get_entity, ['@phpbb.ideas.idea']]
 
     phpbb.ideas.livesearch_controller:
         class: phpbb\ideas\controller\livesearch_controller
         parent: phpbb.ideas.controller.base
+        calls:
+            - [get_entity, ['@phpbb.ideas.ideas']]
 
 # ----- Idea Factory Classes -----
-    phpbb.ideas.manager:
-        class: phpbb\ideas\factory\manager
-        arguments:
-            - '@phpbb.ideas.idea'
-            - '@phpbb.ideas.ideas'
-            - '@phpbb.ideas.vote'
-
     phpbb.ideas.base:
         class: phpbb\ideas\factory\ideas
         arguments:
@@ -95,12 +95,6 @@ services:
     phpbb.ideas.idea:
         class: phpbb\ideas\factory\idea
         parent: phpbb.ideas.base
-        arguments:
-            - '@phpbb.ideas.vote'
-
-    phpbb.ideas.vote:
-        class: phpbb\ideas\factory\vote
-        parent: phpbb.ideas.base
 
     phpbb.ideas.linkhelper:
         class: phpbb\ideas\factory\linkhelper
@@ -119,7 +113,7 @@ services:
         class: phpbb\ideas\cron\prune_orphaned_ideas
         arguments:
             - '@config'
-            - '@phpbb.ideas.manager'
+            - '@phpbb.ideas.ideas'
         calls:
             - [set_name, [cron.task.prune_orphaned_ideas]]
         tags:
