+++ event/listener.php
@@ -14,7 +14,7 @@ use phpbb\auth\auth;
 use phpbb\config\config;
 use phpbb\controller\helper;
 use phpbb\ideas\ext;
-use phpbb\ideas\factory\manager as ideas;
+use phpbb\ideas\factory\idea;
 use phpbb\ideas\factory\linkhelper;
 use phpbb\language\language;
 use phpbb\template\template;
@@ -32,8 +32,8 @@ class listener implements EventSubscriberInterface
 	/* @var helper */
 	protected $helper;
 
-	/* @var ideas */
-	protected $ideas;
+	/* @var idea */
+	protected $idea;
 
 	/** @var language */
 	protected $language;
@@ -54,19 +54,19 @@ class listener implements EventSubscriberInterface
 	 * @param \phpbb\auth\auth                $auth
 	 * @param \phpbb\config\config            $config
 	 * @param \phpbb\controller\helper        $helper
-	 * @param \phpbb\ideas\factory\manager    $ideas
+	 * @param \phpbb\ideas\factory\idea       $idea
 	 * @param \phpbb\language\language        $language
 	 * @param \phpbb\ideas\factory\linkhelper $link_helper
 	 * @param \phpbb\template\template        $template
 	 * @param \phpbb\user                     $user
 	 * @param string                          $php_ext
 	 */
-	public function __construct(auth $auth, config $config, helper $helper, ideas $ideas, language $language, linkhelper $link_helper, template $template, user $user, $php_ext)
+	public function __construct(auth $auth, config $config, helper $helper, idea $idea, language $language, linkhelper $link_helper, template $template, user $user, $php_ext)
 	{
 		$this->auth = $auth;
 		$this->config = $config;
 		$this->helper = $helper;
-		$this->ideas = $ideas;
+		$this->idea = $idea;
 		$this->language = $language;
 		$this->link_helper = $link_helper;
 		$this->template = $template;
@@ -143,7 +143,7 @@ class listener implements EventSubscriberInterface
 			return;
 		}
 
-		$idea = $this->ideas->get_idea_by_topic_id($event['topic_data']['topic_id']);
+		$idea = $this->idea->get_idea_by_topic_id($event['topic_data']['topic_id']);
 
 		if (!$idea)
 		{
@@ -174,7 +174,7 @@ class listener implements EventSubscriberInterface
 		$s_voted_up = $s_voted_down = false;
 		if ($idea['idea_votes_up'] || $idea['idea_votes_down'])
 		{
-			$votes = $this->ideas->get_voters($idea['idea_id']);
+			$votes = $this->idea->get_voters($idea['idea_id']);
 
 			foreach ($votes as $vote)
 			{
@@ -198,9 +198,9 @@ class listener implements EventSubscriberInterface
 			'IDEA_VOTES_DOWN'	=> $idea['idea_votes_down'],
 			'IDEA_POINTS'		=> $points,
 			'IDEA_STATUS_ID'	=> $idea['idea_status'],
-			'IDEA_STATUS_NAME'	=> $this->ideas->get_status_from_id($idea['idea_status']),
+			'IDEA_STATUS_NAME'	=> $this->idea->get_status_from_id($idea['idea_status']),
 
-			'IDEA_DUPLICATE'	=> $idea['duplicate_id'] ? $this->ideas->get_title($idea['duplicate_id']) : '',
+			'IDEA_DUPLICATE'	=> $idea['duplicate_id'] ? $this->idea->get_title($idea['duplicate_id']) : '',
 			'IDEA_RFC'			=> $idea['rfc_link'],
 			'IDEA_TICKET'		=> $idea['ticket_id'],
 			'IDEA_IMPLEMENTED'	=> $idea['implemented_version'],
@@ -341,7 +341,7 @@ class listener implements EventSubscriberInterface
 			return;
 		}
 
-		$this->ideas->submit($event['data']);
+		$this->idea->submit($event['data']);
 
 		// Show users who's posts need approval a special message
 		if (!$this->auth->acl_get('f_noapprove', $event['data']['forum_id']))
@@ -370,8 +370,8 @@ class listener implements EventSubscriberInterface
 			return;
 		}
 
-		$idea = $this->ideas->get_idea_by_topic_id($event['topic_id']);
-		$this->ideas->set_title($idea['idea_id'], $event['post_data']['post_subject']);
+		$idea = $this->idea->get_idea_by_topic_id($event['topic_id']);
+		$this->idea->set_title($idea['idea_id'], $event['post_data']['post_subject']);
 	}
 
 	/**
